+++ common/src/main/java/net/mehvahdjukaar/moonlight/core/network/ClientBoundFinalizeFluidsMessage.java
@@ -1,9 +1,13 @@
 package net.mehvahdjukaar.moonlight.core.network;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.mehvahdjukaar.moonlight.api.misc.DynamicHolder;
+import net.mehvahdjukaar.moonlight.api.misc.HolderReference;
 import net.mehvahdjukaar.moonlight.api.platform.network.Message;
 import net.mehvahdjukaar.moonlight.core.Moonlight;
 import net.mehvahdjukaar.moonlight.core.fluid.SoftFluidInternal;
+import net.minecraft.client.Minecraft;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 
@@ -24,11 +28,13 @@ public class ClientBoundFinalizeFluidsMessage implements Message {
 
     }
 
+    @Environment(EnvType.CLIENT)
     @Override
     public void handle(Context context) {
-        SoftFluidInternal.postInitClient();
+        SoftFluidInternal.postInitClient(Minecraft.getInstance().level.registryAccess());
         //just incase
         DynamicHolder.clearCache();
+        HolderReference.clearCache();
     }
 
     @Override