+++ neoforge/src/main/java/net/mehvahdjukaar/moonlight/neoforge/MoonlightForge.java
@@ -117,7 +117,7 @@ public class MoonlightForge {
 
     @SubscribeEvent
     public static void beforeServerStart(ServerAboutToStartEvent event) {
-        Moonlight.beforeServerStart();
+        Moonlight.beforeServerStart(event.getServer().registryAccess());
     }
 
     @SubscribeEvent